listFiles API Service - Completion
