+++ src/gcs/gcsService.ts
@@ -77,6 +77,8 @@ export class GcsService {
       }
     });
 
+    //const result = await response.json();
+    //console.log(JSON.stringify(result));
     return (await response.json()) as storage_v1.Schema$Objects;
   }
 
@@ -104,7 +106,7 @@ export class GcsService {
     const data = (await requestAPI(
       `api/storage/listFiles?prefix=${prefix}&bucket=${bucket}`
     )) as any;
-    console.log(data);
+    console.log("list files api - result : " , data);
     return data;
   }
